Reset the color in the tear down of robotium tests.
Old Implementation
------------------
The old code did not have any of the desired effects.
I.e. the color and other stuff was not reset.

One reason was the usage of PaintroidApplication.currentTool.getDrawPaint().
This function returns a new instance of a Paint (new Paint(other_paint)).
The stroke width, the stroke cap and the color are of primitive type in Paint.
This means that the new instance has a copy (!) of these values.
Modifying these values for a new Paint instance does not influence the original instance.

New Implementation
------------------
At first I modified getDrawPaint() to not return a copy of Paint and tried
further changes in BaseTool.
Yet none of these changes lead to the desired results.

What works though is to call ColorPickerDialog.getInstance().updateColorChange
directly, which is done now.
For that the updateColorChange function was made public.