Automatize the creation of independent APKs.
Independent APKs are allowed to be installed without affecting the
current Pocket Paint installation.
This is useful for testers to have different versions installed in parallel.
Automating this process makes it more accessible and allows Jenkins
to provide such independent APKs.

Usage
-----
./gradlew -Pindependent assembleDebug

The created Paintroid-debug.apk file has an internal application name
that depends on the current date and time.
The user displayed label is of the form "Paint MMdd HH:mm" in UTC.

Alternatively you can provide an application name yourself:
./gradlew -Pindependent="Paint Design #2" assembleDebug

Implementation
--------------
The application name and the package name are stored as placeholders in
android manifest.
Moreover the test application id is not set manually anymore, since it
is created automatically by the android gradle plugin.