[PAINT-299] Fix adbDisableAnimationsGlobally-Task
For UI-Tests the animations shall be disabled. This is
done via a gradle task adbDisableAnimationsGlobally.
Unfortunately the task is broken, because an exec-task
will only execute one statement. In our case the
two others are skipped. Fixed this circumstance.
Additional clean up unneeded tasks, as the permission
is not needed if the animations are disabled globally.
As all espresso tests used the SystemAnimationsRule,
this can now be safely removed.
Add an adbResetAnimationsGlobally, so the user can
decide to reset it again.