JENKINS-221: Run unit tests
Run local unit tests in separate build step.
Run local unit tests only for debug variant, otherwise
the tests run twice (for debug and release) and so they
are also reported twice.
Merging of coverage reports are not supported out of the
box or via the plugin. So do not report coverage for
the unit tests at the moment.