JENKINS-250 Ensure that build errors are found when running the device tests.
Using "|| true" here would mean that compile errors would not be found.
The original goal of continuing a build for flaky tests is not as important
as detecting compile errors.
Furthermore the build can be continued already when using the ignoreFailures
gradle setting on tests if that really was an issue.