JENKINS-252 Clear the AVD store in the workspace after every build.
The AVDs are recreated for every build anyway.
Thus keeping the AVDs around to clear them on the next build only
increases the harddrive requirements of the slaves.
And since we already had nearly full disks it is better to clean the AVDs.

In the case of Paintroid this safes roughly 820 MB per job.