JENKINS-246 Limit resources used by the Paintroid docker container.
The container is only allowed to use 7 GB of RAM and 3.5 CPUs.
With the -XX:+UnlockExperimentalVMOptions and -XX:+UseCGroupMemoryLimitForHeap
Java Option these settings are also respected in Java 8.
Thus Java 8 uses fewer threads and also memory.

With the limited resources the build jobs take slightly longer than without
restrictions. Yet this allows to have two executors per node in the future
and therefore improving overall performance.
Furthermore this can be seen as necessary precondition for parallelisation:
If there are only few executors there are not enough resources to actually
parallelise things.

The container will only run on executors that provide the label LimitedEmulator.
Nodes with this label should either have just one executor or should not provide
the Emulator label.
This allows older branches that do not restrict the resources to coexist on
Jenkins without negatively affecting each other.

The slaves have 16 GB of RAM and 4 CPUs (8 threads).
These restrictions enable to run two jobs in parallel, reducing the risk
of them affecting each other negatively.