PAINTROID-190 Target API 29: permission to write images to media files
Changed FileIO so that we store with MediaStore library if api > 28.
Otherwise we will save pictures in the Picture folder.
We will not store in Pocket Code Folder anymore.
Changed CreateNewEmptyFile so that we don't have a deprecated
function for api level 29 or higher.

Added Method in MainactivityPresenter so that the permission handling differs
from api version.
Added Method in MainactivityPresenter where we get the mainactivity. Used
for FileIO.
Added Constant in MainactivityConstants for Permission Load_Picture.

Added 3 tests to MainActivityPresenterTest for permission handling.
Rewrote Imagefile tests.