PAINTROID-196
Implemented sharing pictures with Paintroid and the app shows up when the user clicks on any image and tries to share it.
A bitmap with the picture will be created (equal to load image) and when user saves the image a new image is saved.
When the user clicks on the share button the MainActivity gets invoked and identifies, that someone tries to share data with us.
So each time there will be a new instance of the app. Also added new test FileFromOtherApplicationIntegrationTest to test if intent is correctly handled and all of the data is correct.