PAINTROID-188
Implemented that for certain tools the IndeterminateProgressDialog does not show up. Especially for tools that only a really short time (BaseCommand, PointCommand, PathCommand) in order to get rid of the disturbing flickering when drawing. I'm not sure about other Commands if it's necessary because they don't get used as often as the BrushTool.