PAINTROID-122
First approach on how to fix keeping image file size kind of stable after rotation. Highly depends on the format of the image, png images are not really affected but jpg are. As soon as we know which format was loaded/imported/shared we set the given mimetype to compress in the correct format.