Maintenance: Fix crowdin import (#850)
* Fix crowdin import

Deleting incompatible Paintroid/src/main/res/values-kab-rKAB/string.xml (Berber translation) --- error was:

/home/catroid/jenkins_slave/workspace/Paintroid_develop/Paintroid/src/main/res/values-kab-rKAB: Error: Invalid resource directory name 


* Delete combination tw (Twi) with TW (Taiwan)

Incorrect combination of tw (Twi) with TW (Taiwan) translation resource --- lint warning:

Suspicious Language/Region Combination Suspicious language and region combination tw (Twi) with TW (Taiwan): language tw is usually paired with: GH (Ghana) Android uses the letter codes ISO 639-1 for languages, and the letter codes ISO 3166-1 for the region codes. In many cases, the language code and the country where the language is spoken is the same, but it is also often not the case. For example, while 'se' refers to Sweden, where Swedish is spoken, the language code for Swedish is **not** `se` (which refers to the Northern Sami language), the language code is `sv`. And similarly the region code for `sv` is El Salvador. This lint check looks for suspicious language and region combinations, to help catch cases where you've accidentally used the wrong language or region code. Lint knows about the most common regions where a language is spoken, and if a folder combination is not one of these, it is flagged as suspicious. Note however that it may not be an error: you can theoretically have speakers of any language in any region and want to target that with your resources, so this check is aimed at tracking down likely mistakes, not to enforce a specific set of region and language combinations.