PAINTROID-229 Width preview and other settings incorrectly positioned for right to left languages like Arabic
After the changes made the brush tools circle and rectangle is positioned in its place which is left to right when the language of the device is right to left like arabic.

PAINTROID-229 Width preview and other settings incorrectly positioned for right to left languages like Arabic

After the changes made the brush tools circle and rectangle is positioned in its place which is left to right when the language of the device is right to left like arabic. I also added the test for it and ltr is working fine.

PAINTROID-229 The brush tools alignment in right to left languages such as arabic

The strange behavior of cursor on some device while using RTL languages like arabic solved.