PAINTROID-222: Keyboard appears when clicked on color picker. (#864)
* CATROID-229 Width preview and other settings incorrectly positioned for right to left languages like Arabic

After the changes made the brush tools circle and rectangle is positioned in its place which is left to right when the language of the device is right to left like arabic.

* PAINTROID-222 Keyboard appears when clicked on color picker

After the changes made now the keyboard only opens when you are in rgb tab view of color picker and not in hsv or pre tab views. This error used to occur in android 5 to 8 now its been resolved for all versions.