Added a watercolor tool: -Based on the brush tool -Adds an MaskFilter to the paint which immitates watercolor. -Also shows the color in the colorpreview on the bottom right.
Added an AlphaSlider in Colorpicker:
-Instead of only appearing in the HsvColorpicker the slider now appears also in
PreSelectorview. It does not appear in the RGBview.
-When the app is opened from catroid the slider disappears.
-The Slider controls the alpha value of the color. However when the watercolor tool is selected the Slider controls
the strength of the blureffect.

Added watercolor icon.

Added Tests;
-New tests for watercolortool
-Additional tests for colorpickerintegration
-fixed other tests for the new implementation.

Bugs:
(visual)
-For a new ticket, when the watercolor tool is selected and you change the alpha slider in the
colorpicker, you get a new color with a lower alpha value. When you now change to another tool
with a previewcolor (e.g. Brush). The alpha value of the preview color will be the 255(maximum) and
in the colorpicker the value will still be the old value.