PAINTROID-418
Made distinction between different display resolutions (ldpi, mdpi, hdpi, xhdpi, xxhdpi, xxxhdpi) to determine notfication/status bar height. Now when user clicks fullscreen, the bitmap is maxed out as good as possible. Sometimes the behaviour is not optimal, this mostly happens when bitmapheight > surfaceheight, then the scale is not always fully zoomed but I was not able to find a proper solution for this. Also fixed bug where the bitmap was not centered after exiting fullscreen mode, this happened due to the fact that the surfaceheight to determine the middle of the screen was still too small (actual screen size - top-/bottom bar).